requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
requests>=2.31.0
py3Dmol>=2.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
biopython>=1.79
pandas>=1.5.0
plotly>=5.17.0
numpy>=1.24.0
python-dotenv>=1.0.0
